Add Point class Starting of checks in sniffer

+++ src/simulation/tools/SolutionGenerator.java
@@ -11,8 +11,8 @@ public class SolutionGenerator {
 		
 	  	for (BasicSniffer sniffer : this.problem.listOfSniffers) {
 			Random num = new Random();
-			int x = num.nextInt(this.problem.getMaxX());
-			int y = num.nextInt(this.problem.getMaxY());
+			double x = num.nextInt(this.problem.getMaxX());
+			double y = num.nextInt(this.problem.getMaxY());
 			sniffer.setLocation(x, y);
 		}		
 		//return this.problem;
